backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtl-sdr-dashboard"
version = "0.1.0"
description = "RTL-SDR Dashboard Backend - Real-time sensor monitoring and signal analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RTL-SDR Dashboard", email = "rtl@localhost"}
]
keywords = ["rtl-sdr", "sdr", "sensors", "mqtt", "home-assistant", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "paho-mqtt>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.26.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
rtl-dashboard = "app.main:run"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
